Squashed 'gameEngine/' changes from 98debfb..53f7a8a
53f7a8a Merge pull request #8 from Julian52575/game_engine-dev
1a6c3c0 [+] print for collisions
23441b8 [+] Add setentity
dff8406 change compile.sh
f138e55 [fix] make entity
3fcd4ed [fix] Reworked GraphicalCore into Rengine::Rengine
f470538 [WIP] Reworking graphical core into helper class
eaae7fe [fix] Change lua include depending on CMake compiling
ae73fda [add] delay between attack
e5d2ab4 [fix] Dockefile & script && Stupid windows characters in .gitignore
fcba3b5 [fix] upadte position and velocity emplace if doesn't exist
56a9ce3 [fix] if font don't load ,it's catch
de775a1 Merge branch 'game_engine-dev' of github.com:Xekko1629/R-Type into game_engine-dev
983d1fb [fix] warning and [add] parsing JSON better
731c114 [fix] Compiling on docker
1db8f2f [add] Installing .dll and headers
6c32532 [fix] Compiling .dll in windows
3c8bf15 [WIP] Powershell script for windows
9501d56 [fix] Installing conan if not already installed
b8b99d8 [fix] best Maker who use uint32_t entity and regisetry and dump and can update entity
e83cdb3 [~] Change install script
23b96c0 [~] Update lib network to fix bug and remove useless file

git-subtree-dir: gameEngine
git-subtree-split: 53f7a8a231b1969a5844cbb3120d5809d7f2c312

+++ src/System/Animation/Animation.cpp
@@ -5,7 +5,7 @@ System::Animation::Animation(){}
 System::Animation::~Animation(){}
 
 void System::Animation::update(sparse_array<Component::Animation>& animations, sparse_array<Sprite>& sprites, float deltaTime){
-    for (size_t i = 0; i < animations.size() && i < sprites.size(); i++) {
+    for (uint32_t i = 0; i < animations.size() && i < sprites.size(); i++) {
         if (animations[i].has_value() && sprites[i].has_value() && !sprites[i].value().isdefaultTexture){
             animations[i].value().update(deltaTime);
             sprites[i].value().setTextureRect(animations[i].value().getFrame());
